refactor(extract/apple): fold product.go into security-releases.go
